+++ .travis.yml
@@ -1,29 +1,26 @@
----
 sudo: required
 dist: trusty
 
-language: ruby
-
-env:
-  - DB=postgres
-  - DB=mysql
-
-gemfile:
-  - gemfiles/spree_3_1.gemfile
-  - gemfiles/spree_3_2.gemfile
-  - gemfiles/spree_master.gemfile
-
+before_script: "bundle install"
 script:
   - bundle exec rake test_app
   - bundle exec rake spec
+
+env:
+  - DB=mysql
+  - DB=postgres
+
+language: ruby
 
 rvm:
   - 2.3.1
   - 2.2.7
 
-addons:
-  apt:
-    packages:
-    - mysql-server-5.6
-    - mysql-client-core-5.6
-    - mysql-client-5.6
+gemfile:
+  - gemfiles/spree_3_1.gemfile
+  - gemfiles/spree_3_2.gemfile
+  - gemfiles/spree_3_3.gemfile
+  - gemfiles/spree_master.gemfile
+
+before_install:
+  - mysql -u root -e "GRANT ALL ON *.* TO 'travis'@'%';"
