+++ Appraisals
@@ -1,13 +1,22 @@
 appraise 'spree-3-1' do
   gem 'spree', '~> 3.1.0'
+  gem 'spree_auth_devise', '~> 3.1.0'
 end
 
 appraise 'spree-3-2' do
-  gem 'spree', '~> 3.2.0.rc3'
+  gem 'spree', '~> 3.2.0'
+  gem 'spree_auth_devise', '~> 3.2.0'
+  gem 'rails-controller-testing'
+end
+
+appraise 'spree-3-3' do
+  gem 'spree', '~> 3.3.0'
+  gem 'spree_auth_devise', '~> 3.3.0'
   gem 'rails-controller-testing'
 end
 
 appraise 'spree-master' do
   gem 'spree', github: 'spree/spree', branch: 'master'
+  gem 'spree_auth_devise', github: 'spree/spree_auth_devise', branch: 'master'
   gem 'rails-controller-testing'
 end
