+++ Gemfile
@@ -2,6 +2,5 @@ source 'https://rubygems.org'
 
 branch = 'master'
 gem 'spree', github: 'spree/spree', branch: branch
-gem 'spree_auth_devise', github: 'spree/spree_auth_devise', branch: branch
 
 gemspec
