+++ gemfiles/spree_3_1.gemfile
@@ -3,6 +3,6 @@
 source "https://rubygems.org"
 
 gem "spree", "~> 3.1.0"
-gem "spree_auth_devise", :github => "spree/spree_auth_devise", :branch => "master"
+gem "spree_auth_devise", "~> 3.1.0"
 
-gemspec :path => "../"
+gemspec path: "../"
