+++ gemfiles/spree_3_3.gemfile
@@ -2,8 +2,8 @@
 
 source "https://rubygems.org"
 
-gem "spree", "~> 3.2.0"
-gem "spree_auth_devise", "~> 3.2.0"
+gem "spree", "~> 3.3.0"
+gem "spree_auth_devise", "~> 3.3.0"
 gem "rails-controller-testing"
 
 gemspec path: "../"
