+++ gemfiles/spree_master.gemfile
@@ -2,8 +2,8 @@
 
 source "https://rubygems.org"
 
-gem "spree", :github => "spree/spree", :branch => "master"
-gem "spree_auth_devise", :github => "spree/spree_auth_devise", :branch => "master"
+gem "spree", github: "spree/spree", branch: "master"
+gem "spree_auth_devise", github: "spree/spree_auth_devise", branch: "master"
 gem "rails-controller-testing"
 
-gemspec :path => "../"
+gemspec path: "../"
