--- lib/generators/spree/recently_viewed/install/install_generator.rb
@@ -1,16 +1,0 @@
-module Spree
-  module RecentlyViewed
-    module Generators
-      class InstallGenerator < Rails::Generators::Base
-        def self.source_paths
-          paths = self.superclass.source_paths
-          paths.flatten
-        end
-
-        def add_javascripts
-          append_file 'vendor/assets/javascripts/spree/frontend/all.js', "//= require spree/frontend/spree_recently_viewed\n"
-        end
-      end
-    end
-  end
-end
