+++ lib/spree_recently_viewed/version.rb
@@ -9,10 +9,9 @@ module SpreeRecentlyViewed
 
   module VERSION
     MAJOR = 3
-    MINOR = 2
+    MINOR = 3
     TINY  = 0
-    PRE   = 'alpha'.freeze
 
-    STRING = [MAJOR, MINOR, TINY, PRE].compact.join('.')
+    STRING = [MAJOR, MINOR, TINY].compact.join('.')
   end
 end
